Move SIR NN to seperate file; make scaling factors controllable from the config
